Working with files. This program read house prices from a file into a list. And:
a. Determine and print the cost of the most expensive house
b. Determine and print the cost of the cheapest house
c. Determine and print the average cost of the houses
d. Write the prices in reverse order into a file called "dearest_to_cheapest.txt"

Functions created and called.